+++ cosmos/runtime/app.go
@@ -39,7 +39,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/runtime"
 	servertypes "github.com/cosmos/cosmos-sdk/server/types"
 	testdata_pulsar "github.com/cosmos/cosmos-sdk/testutil/testdata/testpb"
-	"github.com/cosmos/cosmos-sdk/types/mempool"
 	"github.com/cosmos/cosmos-sdk/types/module"
 	"github.com/cosmos/cosmos-sdk/x/auth"
 	"github.com/cosmos/cosmos-sdk/x/auth/ante"
@@ -107,10 +106,8 @@ func NewPolarisApp( //nolint:funlen // as defined by the sdk.
 	var (
 		app          = &PolarisApp{}
 		appBuilder   *runtime.AppBuilder
-		ethTxMempool mempool.Mempool = evmmempool.NewEthTxPoolFrom(
-			evmmempool.DefaultPriorityMempool(),
-		)
-		appConfig = depinject.Configs(
+		ethTxMempool = evmmempool.NewPolarisEthereumTxPool()
+		appConfig    = depinject.Configs(
 			AppConfig,
 			depinject.Supply(
 				app.App,
